Fix tests for current package versions.
There are currently two sets of dependency versions for sandman: one set
of pinned versions in `requirements.txt` and another set of unpinned
versions in `setup.py`. This means that installing via `pip` installs
the latest versions of the dependencies, but installing from
`requirements.txt` installs older, pinned versions. Several tests fail
with the latest versions of Flask-Admin and Flask-WTForms.

This patch fixes the breaking tests and installs tests on Travis using
`setup.py` for consistency.